requirements.txt
crawl4ai
browser_cookie3
tenacityaiohttp
aiofiles
lxml
